buffer: Move to a separate package
The module package is meant for interfaces implemented by modules and is
not a catch-all for all utilities.
