+++ dispatcher_test.go
@@ -5,6 +5,7 @@ import (
 	"testing"
 
 	"github.com/emersion/go-message/textproto"
+	"github.com/emersion/maddy/buffer"
 	"github.com/emersion/maddy/module"
 )
 
@@ -247,7 +248,7 @@ func TestDispatcher_PerRcptReject(t *testing.T) {
 	if err := delivery.AddRcpt("rcpt1@example.com"); err != nil {
 		t.Fatalf("unexpected AddRcpt err for %s: %v", "rcpt1@example.com", err)
 	}
-	if err := delivery.Body(textproto.Header{}, module.MemoryBuffer{Slice: []byte("foobar")}); err != nil {
+	if err := delivery.Body(textproto.Header{}, buffer.MemoryBuffer{Slice: []byte("foobar")}); err != nil {
 		t.Fatalf("unexpected Body err: %v", err)
 	}
 	if err := delivery.Commit(); err != nil {
