+++ dummy.go
@@ -2,6 +2,7 @@ package maddy
 
 import (
 	"github.com/emersion/go-message/textproto"
+	"github.com/emersion/maddy/buffer"
 	"github.com/emersion/maddy/config"
 	"github.com/emersion/maddy/module"
 )
@@ -36,7 +37,7 @@ func (dd dummyDelivery) AddRcpt(to string) error {
 	return nil
 }
 
-func (dd dummyDelivery) Body(header textproto.Header, body module.Buffer) error {
+func (dd dummyDelivery) Body(header textproto.Header, body buffer.Buffer) error {
 	return nil
 }
