+++ module/delivery_target.go
@@ -2,6 +2,7 @@ package module
 
 import (
 	"github.com/emersion/go-message/textproto"
+	"github.com/emersion/maddy/buffer"
 )
 
 type DeliveryTarget interface {
@@ -10,7 +11,7 @@ type DeliveryTarget interface {
 
 type Delivery interface {
 	AddRcpt(rcptTo string) error
-	Body(header textproto.Header, body Buffer) error
+	Body(header textproto.Header, body buffer.Buffer) error
 
 	Abort() error
 	Commit() error
